feat: add test with example trace
This test gives a concrete example of how the structures can be put
together to represent the execution of a program. The actual logic that
is tested is contrived: it's just checkinging that the serialized
message is of a certain length, but this could be flaky, as it depends
on the behavior of the serde dependency.